moved main files around
